# docs-site/contracts/product-http.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WuKongIM HTTP API",
    "version": "3.0.0-beta-snapshot",
    "description": "The 48 core WuKongIM HTTP API JSON operations, plus the plugin-defined ANY route in x-wukongim-extension-routes. Routes have no built-in authentication.",
    "x-i18n": {
      "zh": {
        "title": "WuKongIM HTTP API",
        "description": "与当前运行时的 48 个核心 WuKongIM HTTP API JSON 操作对齐；插件定义的 ANY 路由见 x-wukongim-extension-routes。这些路由没有内建鉴权。"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5001",
      "description": "Local single-node cluster example",
      "x-i18n": {
        "zh": {
          "description": "本地单节点集群示例"
        }
      }
    }
  ],
...
  "security": [],
  "paths": {
...
    "/message/eventsync": {
      "post": {
        "operationId": "syncMessageEvents",
        "summary": "Synchronize current message event projections",
        "description": "Reads current durable event projections after a sequence cursor, with filtering after a bounded read; not a replayable event log.",
        "x-i18n": {
          "zh": {
            "summary": "同步消息事件当前投影",
            "description": "按序号游标读取当前持久化事件投影，不重放完整事件日志。在有界读取后过滤私有事件。调用方身份与 include_private 没有鉴权，必须由受信后端保护。"
          }
        },
        "tags": [
          "Messages"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/MessageEventSync"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MessageEventSyncOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
...
    "requestBodies": {
...
      "MessageEventSync": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageEventSyncRequest"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "CompatibilityBadRequest": {
        "description": "Input, validation, dependency, or downstream error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CompatibilityError"
            }
          }
        }
      },
      "MaintenanceUnavailable": {
        "description": "Controller restore maintenance is active",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MaintenanceError"
            }
          }
        }
      },
...
      "MessageEventSyncOK": {
        "description": "Current durable event projections",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageEventSyncResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "CompatibilityError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "msg",
          "status"
        ],
        "properties": {
          "msg": {
            "type": "string",
            "description": "Human-readable and not stable for machine matching."
          },
          "status": {
            "type": "integer",
            "const": 400
          }
        }
      },
      "MaintenanceError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "const": "maintenance"
          },
          "message": {
            "type": "string",
            "const": "restore maintenance is active"
          }
        }
      },
...
      "MessageEventSyncRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "channel_id",
          "channel_type",
          "client_msg_no"
        ],
        "properties": {
          "channel_id": {
            "type": "string",
            "pattern": "\\S"
          },
          "channel_type": {
            "type": "integer",
            "format": "uint8",
            "minimum": 1,
            "maximum": 255
          },
          "from_uid": {
            "type": "string"
          },
          "client_msg_no": {
            "type": "string",
            "pattern": "\\S"
          },
          "event_key": {
            "type": "string"
          },
          "from_msg_event_seq": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0,
            "default": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 0,
            "default": 0
          },
          "include_private": {
            "type": "integer",
            "format": "uint8",
            "minimum": 0,
            "maximum": 255,
            "default": 0
          }
        }
      },
      "SyncedMessageEvent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "msg_event_seq",
          "event_id",
          "event_key",
          "event_type",
          "visibility",
          "occurred_at"
        ],
        "properties": {
          "msg_event_seq": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "event_id": {
            "type": "string"
          },
          "event_key": {
            "type": "string"
          },
          "event_type": {
            "type": "string"
          },
          "visibility": {
            "type": "string"
          },
          "occurred_at": {
            "type": "integer",
            "format": "int64"
          },
          "payload": {}
        }
      },
      "MessageEventSyncData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "client_msg_no",
          "from_msg_event_seq",
          "next_msg_event_seq",
          "more",
          "events",
          "filtered_by_event_key"
        ],
        "properties": {
          "client_msg_no": {
            "type": "string"
          },
          "from_msg_event_seq": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "next_msg_event_seq": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "more": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "events": {
            "type": "array",
            "maxItems": 2000,
            "items": {
              "$ref": "#/components/schemas/SyncedMessageEvent"
            }
          },
          "filtered_by_event_key": {
            "type": "string"
          }
        }
      },
      "MessageEventSyncResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "const": 200
          },
          "data": {
            "$ref": "#/components/schemas/MessageEventSyncData"
          }
        }
      },
...
    }
  }
}